优化缓存设计，优化Agent设计
