+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "amrita"
-version = "0.3.3.2"
+version = "0.3.4"
 description = "A powerful bot framework powered by NoneBot2"
 readme = "README.md"
 license = "GPL-3.0-or-later"
